Namespace changes and naming cleanup
