+++ src/Umbraco.Cms.Api.Management/Controllers/Document/Version/RollbackDocumentVersionController.cs
@@ -6,14 +6,14 @@ using Umbraco.Cms.Core.Security;
 using Umbraco.Cms.Core.Services;
 using Umbraco.Cms.Core.Services.OperationStatus;
 
-namespace Umbraco.Cms.Api.Management.Controllers.Document.Versions;
+namespace Umbraco.Cms.Api.Management.Controllers.Document.Version;
 
 [ApiVersion("1.0")]
-public class UpdatePreventCleanupDocumentVersionController : DocumentVersionControllerBase
+public class RollbackDocumentVersionController : DocumentVersionControllerBase
 {
     private readonly IBackOfficeSecurityAccessor _backOfficeSecurityAccessor;
 
-    public UpdatePreventCleanupDocumentVersionController(
+    public RollbackDocumentVersionController(
         IContentVersionService contentVersionService,
         IBackOfficeSecurityAccessor backOfficeSecurityAccessor)
         : base(contentVersionService)
@@ -22,14 +22,14 @@ public class UpdatePreventCleanupDocumentVersionController : DocumentVersionCont
     }
 
     [MapToApiVersion("1.0")]
-    [HttpPut("{id:guid}/preventCleanup")]
+    [HttpPost("{id:guid}/rollback")]
     [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status404NotFound)]
     [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status400BadRequest)]
-    public async Task<IActionResult> Set(Guid id, bool preventCleanup)
+    public async Task<IActionResult> Rollback(Guid id, string? culture)
     {
         Attempt<ContentVersionOperationStatus> attempt =
-            await ContentVersionService.SetPreventCleanupAsync(id, preventCleanup, CurrentUserKey(_backOfficeSecurityAccessor));
+            await ContentVersionService.RollBackAsync(id, culture, CurrentUserKey(_backOfficeSecurityAccessor));
 
         return attempt.Success is true
             ? Ok()
